Added GameService tests

+++ gameStore/src/main/java/group15/gameStore/service/GameService.java
@@ -119,7 +119,7 @@ public class GameService {
         if (gameToDelete == null || employee == null) {
             throw new GameStoreException(HttpStatus.BAD_REQUEST, "Invalid game creation request: missing attributes");
         }
-        if (!employee.isIsManager() && !gameToDelete.isIsApproved()) {
+        if (!employee.isIsManager() || !gameToDelete.isIsApproved()) {
             throw new GameStoreException(HttpStatus.UNAUTHORIZED, "The deletion of the game has not been approved by the manager");
         }
         //Search game
@@ -127,6 +127,9 @@ public class GameService {
             throw new GameStoreException(HttpStatus.NOT_FOUND, String.format("The employee '%s' that made the request does not exist", employee.getUsername()));
         }
         Game game = gameRepo.findGameByGameID(gameToDelete.getGameID());
+        if (game == null) {
+            throw new GameStoreException(HttpStatus.NOT_FOUND, "The game to delete does not exist");
+        }
         gameRepo.delete(game);
     }
